Export Peers instead of Endpoints
